Player now passes through clouds and kites from the sides and bottom
This change may make the game a bit easier, but after playing a over the years, I found that not being able to pass through clouds just didn't feel right.

Had I thought of this before essentially finishing the tilemap, I would have added this functionality to a few more "platforms", but I think just doing it for clouds and kites will suffice.